+++ src/lib/types/main.ts
@@ -9,3 +9,10 @@ export type TextColorType = {
   content: string;
   color: 'yellow' | 'white';
 };
+
+export interface ActivityType {
+  img: string;
+  navKor: string;
+  navEng: string;
+  description: TextColorType[];
+}
